+++ webpack.config.common.js
@@ -113,6 +113,7 @@ module.exports = {
     alias: {
       Components: path.resolve(srcScript, 'components'),
       I18n: path.resolve(SRC_PATH, 'i18n'),
+      Modals: path.resolve(srcScript, 'page/modals'),
       Resource: path.resolve(ROOT_PATH, 'resource'),
       Util: path.resolve(srcScript, 'util'),
     },
